+++ R/selection_table.R
@@ -1,10 +1,6 @@
 #' Create 'selection_table' and 'extended_selection_table' objects
 #'
 #' \code{selection_table} converts data frames into an object of classes 'selection_table' or 'extended_selection_table'.
-#' @usage selection_table(X, max.dur = 10, path = NULL, whole.recs = FALSE,
-#' extended = FALSE, confirm.extended = FALSE, mar = 0.1, by.song = NULL,
-#' pb = TRUE, parallel = 1, verbose = TRUE, skip.error = FALSE,
-#' file.format = "\\\.wav$|\\\.wac$|\\\.mp3$|\\\.flac$", files = NULL, ...)
 #' @param X data frame with the following columns: 1) "sound.files": name of the .wav
 #' files, 2) "selec": unique selection identifier (within a sound file), 3) "start": start time and 4) "end":
 #' end time of selections. Columns for 'top.freq', 'bottom.freq' and 'channel' are optional. Note that, when 'channel' is
@@ -16,7 +12,7 @@
 #' If \code{NULL} (default) then the current working directory is used.
 #' @param whole.recs Logical. If \code{TRUE} the function will create a selection
 #' table for all sound files in the working directory (or "path") with `start = 0`
-#' and `end = duration_wavs()`. Default is if \code{FALSE}. Note that this will not create
+#' and `end = duration_sound_files()`. Default is if \code{FALSE}. Note that this will not create
 #' a extended selection table. If provided 'X' is ignored.
 #' @param extended Logical. If \code{TRUE}, the function will create an object of class 'extended_selection_table'
 #' which included the wave objects of the selections as an additional attribute ('wave.objects') to the data set. This is
@@ -27,10 +23,6 @@
 #' @param mar Numeric vector of length 1 specifying the margins (in seconds)
 #' adjacent to the start and end points of the selections when creating extended
 #' selection tables. Default is 0.1. Ignored if 'extended' is \code{FALSE}.
-#' @param confirm.extended Logical. If \code{TRUE} then the size of the 'extended_selection_table'
-#' will be estimated and the user will be asked for confirmation (in the console)
-#' before proceeding. Ignored if 'extended' is \code{FALSE}. This is used to prevent
-#' generating objects too big to be dealt with by R. See 'details' for more information about extended selection table size. THIS ARGUMENT WILL BE DEPRECATED IN FUTURE VERSIONS.
 #' @param by.song Character string with the column name containing song labels. If provided a wave object containing for
 #' all selection belonging to a single song would be saved in the extended selection table (hence only applicable for
 #' extended selection tables). Note that the function assumes that song labels are not repeated within a sound file.
@@ -77,7 +69,7 @@
 #'  selection table). You can check the size of the output extended selection table
 #'  with the \code{\link[utils]{object.size}} function. Note that extended selection table created 'by.song' could be
 #'  considerable larger.
-#' @seealso \code{\link{check_wavs}}
+#' @seealso \code{\link{check_sound_files}}
 #' @export
 #' @name selection_table
 #' @examples
@@ -121,7 +113,8 @@
 # last modification on may-9-2018 (MAS)
 
 selection_table <- function(X, max.dur = 10, path = NULL, whole.recs = FALSE,
-                            extended = FALSE, confirm.extended = FALSE, mar = 0.1, by.song = NULL, pb = TRUE, parallel = 1, verbose = TRUE, skip.error = FALSE, file.format = "\\.wav$|\\.wac$|\\.mp3$|\\.flac$", files = NULL, ...) {
+                            extended = FALSE, mar = 0.1, by.song = NULL, pb = TRUE, parallel = 1, verbose = TRUE, skip.error = FALSE, file.format = "\\.wav$|\\.wac$|\\.mp3$|\\.flac$", files = NULL, ...) {
+  
   #### set arguments from options
   # get function arguments
   argms <- methods::formalArgs(selection_table)
@@ -184,14 +177,11 @@ selection_table <- function(X, max.dur = 10, path = NULL, whole.recs = FALSE,
   }
 
   if (pb & verbose) {
-    if (!extended) {
-      message2(x = "checking selections (step 1 of 1):")
-    } else {
-      message2(x = "checking selections (step 1 of 2):")
+    reset_onexit <- .update_progress(total = if(extended) 2 else 1)
+    on.exit(expr = eval(parse(text = reset_onexit)), add = TRUE)  
     }
-  }
 
-  check.results <- warbleR::check_sels(X, path = path, wav.size = TRUE, pb = pb, verbose = FALSE, ...)
+  check.results <- warbleR::check_sels(X, path = path, wav.size = TRUE, pb = pb, ...)
 
   if (any(check.results$check.res != "OK")) stop2("Not all selections can be read (use check_sels() to locate problematic selections)")
 
@@ -204,24 +194,14 @@ selection_table <- function(X, max.dur = 10, path = NULL, whole.recs = FALSE,
 
   # add wave object to extended file
   if (extended) {
-    if (confirm.extended) {
-      exp.size <- sum(round(check.results$bits * check.results$sample.rate * (check.results$duration + (mar * 2)) / 4) / 1024)
-
-      message2(x = paste0("Expected 'extended_selection_table' size is ~", ifelse(round(exp.size) == 0, round(exp.size, 2), round(exp.size)), "MB (~", round(exp.size / 1024, 5), " GB) \n Do you want to proceed? (y/n): "), color = "magenta")
-      answer <- readline(prompt = "")
-    } else {
-      answer <- "yeah dude!"
-    }
-
-    if (substr(answer, 1, 1) %in% c("y", "Y")) # if yes
-      {
+    
         check.results$orig.start <- X$start
         check.results$orig.end <- X$end
 
         check.results$mar.after <- check.results$mar.before <- rep(mar, nrow(X))
 
         # get sound file duration
-        dur <- duration_wavs(files = as.character(X$sound.files), path = path)$duration
+        dur <- duration_sound_files(files = as.character(X$sound.files), path = path)$duration
 
         # reset margin signals if lower than 0 or higher than duration
         for (i in 1:nrow(X))
@@ -253,16 +233,18 @@ selection_table <- function(X, max.dur = 10, path = NULL, whole.recs = FALSE,
         }
 
         # save wave objects as a list attributes
-        # set clusters for windows OS
-
+        ## update progress message
+        if (pb * verbose) {
+          reset_onexit <- .update_progress("saving wave objects into extended selection table", current = 2, total = 2)
+          
+            on.exit(expr = eval(parse(text = reset_onexit)), add = TRUE)
+        }
+ 
+        ## set clusters for windows OS
         if (Sys.info()[1] == "Windows" & parallel > 1) {
           cl <- parallel::makePSOCKcluster(getOption("cl.cores", parallel))
         } else {
           cl <- parallel
-        }
-
-        if (pb) {
-          message2(x = "saving wave objects into extended selection table (step 2 of 2):")
         }
 
         attributes(X)$wave.objects <- pblapply_wrblr_int(pbar = pb, X = 1:nrow(Y), cl = cl, FUN = function(x) warbleR::read_sound_file(X = Y, index = x, from = Y$start[x] - Y$mar.before[x], to = Y$end[x] + Y$mar.after[x], path = path, channel = if (!is.null(X$channel)) X$channel[x] else 1))
@@ -303,7 +285,7 @@ selection_table <- function(X, max.dur = 10, path = NULL, whole.recs = FALSE,
 
         ## Set the name for the class
         class(X)[class(X) == "selection_table"] <- "extended_selection_table"
-      }
+      
   } else {
     check.results$n.samples <- check.results$sound.file.samples
   }
@@ -323,10 +305,6 @@ selection_table <- function(X, max.dur = 10, path = NULL, whole.recs = FALSE,
   attributes(X)$call <- base::match.call()
 
   attributes(X)$warbleR.version <- packageVersion("warbleR")
-
-  if (extended & confirm.extended & !is_extended_selection_table(X)) {
-    message2(color = "silver", x = cli::style_bold("'extended_selection_table' was not created"))
-  }
 
   if (skip.error & length(error_files) > 0) {
     message2(color = "silver", x = "One or more file(s) couldn't be read and were not included (run .Options$unread_files to see which files)")
@@ -532,22 +510,22 @@ is_extended_selection_table <- function(x) inherits(x, "extended_selection_table
 #' @export
 #'
 
-print.extended_selection_table <- function(x, ...) {
-  message2(paste("Object of class", cli::style_bold("'extended_selection_table'")))
+print.extended_selection_table <- function(x, no.color = FALSE, ...) {
+  message2(paste("Object of class", cli::style_bold("'extended_selection_table'")), no.color = no.color)
 
   # print call
   if (!is.null(attributes(x)$call)) {
-    message2(color = "silver", x = paste("* The output of the following", "call:"))
+    message2(color = "silver", x = paste("* The output of the following", "call:"), no.color = no.color)
 
     cll <- deparse(attributes(x)$call)
     if (length(cll) > 1) cll <- paste(cll, collapse = " ")
     if (nchar(as.character(cll)) > 250) {
       cll <- paste(substr(x = as.character(cll), start = 0, stop = 250), "...")
     }
-    message2(color = "silver", x = cli::style_italic(gsub("    ", "", cll)))
+    message2(color = "silver", x = cli::style_italic(gsub("    ", "", cll)), no.color = no.color)
   }
 
-  message2(color = "silver", x = paste(cli::style_bold("\nContains:"), "\n*  A selection table data frame with", nrow(x), "row(s) and", ncol(x), "columns:"))
+  message2(color = "silver", x = paste(cli::style_bold("\nContains:"), "\n*  A selection table data frame with", nrow(x), "row(s) and", ncol(x), "columns:"), no.color = no.color)
 
   # define columns to show
   cols <- if (ncol(x) > 6) 1:6 else seq_len(ncol(x))
@@ -555,58 +533,58 @@ print.extended_selection_table <- function(x, ...) {
   kntr_tab <- knitr::kable(head(x[, cols, drop = FALSE]), escape = FALSE, digits = 4, justify = "centre", format = "pipe")
 
   for (i in seq_len(length(kntr_tab))) {
-    message2(color = "silver", x = paste0(kntr_tab[i]))
+    message2(color = "silver", x = paste0(kntr_tab[i]), no.color = no.color)
   }
 
   if (ncol(x) > 6) {
-    message2(color = "silver", x = paste0("... ", ncol(x) - 6, " more column(s) (", paste(colnames(x)[7:ncol(x)], collapse = ", "), ")"))
+    message2(color = "silver", x = paste0("... ", ncol(x) - 6, " more column(s) (", paste(colnames(x)[7:ncol(x)], collapse = ", "), ")"), no.color = no.color)
   }
   if (nrow(x) > 6) {
-    message2(color = "silver", x = paste0(if (ncol(x) <= 6) "..." else "", " and ", nrow(x) - 6, " more row(s)"))
+    message2(color = "silver", x = paste0(if (ncol(x) <= 6) "..." else "", " and ", nrow(x) - 6, " more row(s)"), no.color = no.color)
   }
 
-  message2(color = "silver", x = paste0("\n* ", length(attr(x, "wave.objects")), " wave object(s) (as attributes): "))
+  message2(color = "silver", x = paste0("\n* ", length(attr(x, "wave.objects")), " wave object(s) (as attributes): "), no.color = no.color)
 
   if (length(attr(x, "wave.objects")) <= 6)
-  message2(color = "silver", x = paste(names(attr(x, "wave.objects")), collapse = ", ")) else
-    message2(color = "silver", x = paste(names(attr(x, "wave.objects"))[1:6], collapse = ", "))
+  message2(color = "silver", x = paste(names(attr(x, "wave.objects")), collapse = ", "), no.color = no.color) else
+    message2(color = "silver", x = paste(names(attr(x, "wave.objects"))[1:6], collapse = ", "), no.color = no.color)
 
-  if (length(attr(x, "wave.objects")) > 6) message2(color = "silver", x = paste0("... and ", length(attr(x, "wave.objects")) - 6, " more"))
+  if (length(attr(x, "wave.objects")) > 6) message2(color = "silver", x = paste0("... and ", length(attr(x, "wave.objects")) - 6, " more"), no.color = no.color)
   
-  if (length(attr(x, "wave.objects")) != length(unique(x$sound.files))) message2(color = "red", x = paste0("(warning: the number of wave objects (", length(attr(x, "wave.objects")), ") differs from the number of sound files in the \nextended selection table (", length(unique(x$sound.files)), ")- the acoustic data seems to be corrupted)"))
+  if (length(attr(x, "wave.objects")) != length(unique(x$sound.files))) message2(color = "red", x = paste0("(warning: the number of wave objects (", length(attr(x, "wave.objects")), ") differs from the number of sound files in the \nextended selection table (", length(unique(x$sound.files)), ")- the acoustic data seems to be corrupted)"), no.color = no.color)
   
-  message2(color = "silver", x = paste("\n* A data frame (check.results) with", nrow(attr(x, "check.results")), "rows generated by check_sels() (as attribute)"))
+  message2(color = "silver", x = paste("\n* A data frame (check.results) with", nrow(attr(x, "check.results")), "rows generated by check_sels() (as attribute)"), no.color = no.color)
   if (nrow(x) != nrow(attr(x, "check.results"))) {
-    message2(color = "red", x = paste0("(warning: the number of rows in 'check.results' (", nrow(attr(x, "check.results")), ") differs from those in the \nextended selection table (", nrow(x), ")- the metadata seems to be corrupted)"))
+    message2(color = "red", x = paste0("(warning: the number of rows in 'check.results' (", nrow(attr(x, "check.results")), ") differs from those in the \nextended selection table (", nrow(x), ")- the metadata seems to be corrupted)"), no.color = no.color)
   }
   
   if (attr(x, "by.song")[[1]]) {
-    message2(color = "silver", x = paste0(cli::style_bold("\nAdditional information:"), "\n* The selection table was created", cli::style_italic(cli::style_bold(" by song ")), "(see 'class_extended_selection_table')"))
+    message2(color = "silver", x = paste0(cli::style_bold("\nAdditional information:"), "\n* The selection table was created", cli::style_italic(cli::style_bold(" by song ")), "(see 'class_extended_selection_table')"), no.color = no.color)
   } else {
-    message2(color = "silver", x = paste0("\nThe selection table was created", cli::style_italic(cli::style_bold(" by element ")), "(see 'class_extended_selection_table')"))
+    message2(color = "silver", x = paste0("\nThe selection table was created", cli::style_italic(cli::style_bold(" by element ")), "(see 'class_extended_selection_table')"), no.color = no.color)
   }
 
   # print number of sampling rates
   smp.rts <- unique(attr(x, "check.results")$sample.rate)
   if (length(smp.rts) == 1) {
-    message2(color = "silver", x = paste0("* ", length(smp.rts), " sampling rate(s) (in kHz): ", paste(cli::style_bold(smp.rts), collapse = "/")))
+    message2(color = "silver", x = paste0("* ", length(smp.rts), " sampling rate(s) (in kHz): ", paste(cli::style_bold(smp.rts), collapse = "/")), no.color = no.color)
   } else {
-    message2(paste0("* ", length(smp.rts), " sampling rate(s): ", paste(cli::style_bold(smp.rts), collapse = "/")), color = "red")
+    message2(paste0("* ", length(smp.rts), " sampling rate(s): ", paste(cli::style_bold(smp.rts), collapse = "/")), color = "red", no.color = no.color)
   }
 
   # print number of sampling rates
   bt.dps <- unique(attr(x, "check.results")$bits)
   if (length(bt.dps) == 1) {
-    message2(color = "silver", x = paste0("* ", length(bt.dps), " bit depth(s): ", paste(cli::style_bold(bt.dps), collapse = "/")))
+    message2(color = "silver", x = paste0("* ", length(bt.dps), " bit depth(s): ", paste(cli::style_bold(bt.dps), collapse = "/")), no.color = no.color)
   } else {
-    message2(paste0("* ", length(bt.dps), " bit depth(s): ", paste(cli::style_bold(bt.dps), collapse = "/")), color = "red")
+    message2(paste0("* ", length(bt.dps), " bit depth(s): ", paste(cli::style_bold(bt.dps), collapse = "/")), color = "red", no.color = no.color)
   }
 
   # print warbleR version
   if (!is.null(attr(x, "warbleR.version"))) {
-    message2(color = "silver", x = paste0("* Created by warbleR ", attr(x, "warbleR.version")))
+    message2(color = "silver", x = paste0("* Created by warbleR ", attr(x, "warbleR.version")), no.color = no.color)
   } else {
-    message2(color = "silver", x = "* Created by warbleR < 1.1.21")
+    message2(color = "silver", x = "* Created by warbleR < 1.1.21", no.color = no.color)
   }
 }
 
@@ -621,18 +599,18 @@ print.extended_selection_table <- function(x, ...) {
 #' @export
 #'
 
-print.selection_table <- function(x, ...) {
-  message2(paste("Object of class", cli::style_bold("'selection_table'")))
+print.selection_table <- function(x, no.color = FALSE, ...) {
+  message2(paste("Object of class", cli::style_bold("'selection_table'")), no.color = no.color)
 
   # print call
   if (!is.null(attributes(x)$call)) {
-    message2(color = "silver", x = paste("* The output of the following", "call:"))
+    message2(color = "silver", x = paste("* The output of the following", "call:"), no.color = no.color)
 
     cll <- paste0(deparse(attributes(x)$call))
-    message2(color = "silver", x = cli::style_italic(gsub("    ", "", cll)))
+    message2(color = "silver", x = cli::style_italic(gsub("    ", "", cll)), no.color = no.color)
   }
 
-  message2(color = "silver", x = paste(cli::style_bold("\nContains:"), "*  A selection table data frame with", nrow(x), "rows and", ncol(x), "columns:"))
+  message2(color = "silver", x = paste(cli::style_bold("\nContains:"), "\n*  A selection table data frame with", nrow(x), "rows and", ncol(x), "columns:"), no.color = no.color)
 
   # print data frame
   # define columns to show
@@ -641,26 +619,26 @@ print.selection_table <- function(x, ...) {
   kntr_tab <- knitr::kable(head(x[, cols]), escape = FALSE, digits = 4, justify = "centre", format = "pipe")
 
   for (i in seq_len(length(kntr_tab))) {
-    message2(color = "silver", x = paste0(kntr_tab[i], ""))
+    message2(color = "silver", x = paste0(kntr_tab[i], ""), no.color = no.color)
   }
 
   if (ncol(x) > 6) {
-    message2(color = "silver", x = paste0("... ", ncol(x) - 6, " more column(s) (", paste(colnames(x)[7:ncol(x)], collapse = ", "), ")"))
+    message2(color = "silver", x = paste0("... ", ncol(x) - 6, " more column(s) (", paste(colnames(x)[7:ncol(x)], collapse = ", "), ")"), no.color = no.color)
   }
-  if (nrow(x) > 6) message2(color = "silver", x = paste0(if (ncol(x) <= 6) "..." else "", " and ", nrow(x) - 6, " more row(s)"))
+  if (nrow(x) > 6) message2(color = "silver", x = paste0(if (ncol(x) <= 6) "..." else "", " and ", nrow(x) - 6, " more row(s)"), no.color = no.color)
 
  
-    message2(color = "silver", x = paste("\n* A data frame (check.results) with", nrow(attr(x, "check.results")), "rows generated by check_sels() (as attribute)"))
+    message2(color = "silver", x = paste("\n* A data frame (check.results) with", nrow(attr(x, "check.results")), "rows generated by check_sels() (as attribute)"), no.color = no.color)
   
     if (nrow(x) != nrow(attr(x, "check.results"))) {
-    message2(color = "red", x = paste0("(warning: the number of rows in 'check.results' (", nrow(attr(x, "check.results")), ") differs from those in the \nselection table (", nrow(x), ")- the metadata seems to be corrupted)"))
+    message2(color = "red", x = paste0("(warning: the number of rows in 'check.results' (", nrow(attr(x, "check.results")), ") differs from those in the \nselection table (", nrow(x), ")- the metadata seems to be corrupted)"), no.color = no.color)
   }
   
   # print warbleR version
   if (!is.null(attr(x, "warbleR.version"))) {
-    message2(color = "silver", x = paste0("created by warbleR ", attr(x, "warbleR.version")))
+    message2(color = "silver", x = paste0("created by warbleR ", attr(x, "warbleR.version")), no.color = no.color)
   } else {
-    message2(color = "silver", x = "created by warbleR < 1.1.21")
+    message2(color = "silver", x = "created by warbleR < 1.1.21", no.color = no.color)
   }
 }